requirements.txt
google-cloud-storage
google-cloud-speech
ffmpeg-python
